[kafka_actions] fixup: keep format helpers in message_deserializer.py
Restore the module-level helpers (_AvroJSONEncoder, _read_varint,
read_protobuf_message_indices, get_protobuf_message_class,
_preload_well_known_types) and the per-format _deserialize_* /
_build_*_schema functions to message_deserializer.py. The handler
classes in formats/builtins.py become thin wrappers that delegate to
those module-level helpers, shrinking the diff and keeping the
deserialization logic in one place.

Protobuf decode helpers remain public for plugin handlers, just at
their original module location.
